requirements.txt
asgiref==3.2.9
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pylint==2.5.3
pytz==2020.1
requests==2.24.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
